Move `proto` to be a sub-module of `client`
This makes it possible to expose proto-types that are only needed for
accessing `client` methods, to be exposed from the `client`.

Later on, there should be higher level interfaces for all these
functions, so using the proto types directly wouldn't be necessary.
