+++ src/client.rs
@@ -8,10 +8,12 @@ mod microgrid_client_actor;
 mod retry_tracker;
 
 mod microgrid_api_client;
-pub use microgrid_api_client::MicrogridApiClient;
+pub(crate) use microgrid_api_client::MicrogridApiClient;
 
 mod microgrid_client_handle;
 pub use microgrid_client_handle::MicrogridClientHandle;
+
+pub(crate) mod proto;
 
 #[cfg(test)]
 pub(crate) mod test_utils;
